requirements.txt
fastapi
uvicorn
mistralai
pymongo
python-multipart
PyMuPDF
python-docx
python-dotenv
websockets
google-cloud-texttospeech
google-cloud-speech
